TGP-2893: Begin moving files
