+++ app/navigation/Navigation.scala
@@ -23,6 +23,7 @@ import models._
 import models.ott.{CategorisationInfo, CategoryAssessment}
 import pages._
 import pages.categorisation.CategoryGuidancePage
+import pages.download.RequestDataPage
 import play.api.mvc.Call
 import queries.{CategorisationDetailsQuery, LongerCategorisationDetailsQuery, LongerCommodityQuery}
 import services.CategorisationService
