+++ conf/app.routes
@@ -3,6 +3,7 @@
 ->          /                                                                                       app.profile.Routes
 ->          /                                                                                       app.problem.Routes
 ->          /                                                                                       app.categorisation.Routes
+->          /                                                                                       app.download.Routes
 ->          /hmrc-frontend                                                                          hmrcfrontend.Routes
 
 GET         /                                                                                       controllers.IndexController.onPageLoad()
@@ -208,13 +209,6 @@ GET         /update-record/:recordId/withdraw-request/reason                    
 POST        /update-record/:recordId/withdraw-request/reason                                        controllers.ReasonForWithdrawAdviceController.onSubmit(recordId: String)
 
 GET         /download-data/request-received                                                         controllers.DownloadRequestSuccessController.onPageLoad()
-
-
-GET         /download/tgp-record-files                                                              controllers.FileManagementController.onPageLoad()
-
-GET         /download-data/request-file                                                             controllers.RequestDataController.onPageLoad()
-POST        /download-data/request-file                                                             controllers.RequestDataController.onSubmit()
-
 
 # this url must be the last url on this page so it becomes the default if no other urls match
 GET         /*path                                                                                  controllers.PageNotFoundController.onPageLoad(path: String)
